+++ History.md
@@ -5,6 +5,7 @@
    - `source` is now the CSS string
    - `filename` is now the optional filename
  * changed default `options.position` value to `true`
+ * remove comments from properties and values
 
 1.7.0 / 2013-12-21
 ==================