+++ History.md
@@ -1,6 +1,7 @@
 HEAD
 ==================
 
+ * add support for `@custom-media`
  * allow commas inside selector functions
  * allow empty property values
  * changed default `options.position` value to `true`